Refactor constant evaluation
